Refactor main bag.js to use a jobs data factory mapFilesToJobs, move geometry tools tests to separate test file
