+++ .gitignore
@@ -1,3 +1,4 @@
 /test/BAG_Amstelveen_2011feb01.zip
+/coverage
 /test/buildings.ndjson
 /test/unzip
