+++ test/bagTest.js
@@ -13,7 +13,6 @@ var assert = chai.assert;
 var expect = chai.expect;
 
 var bag = require('../bag.js');
-var geometryTools = require('../helpers/geometrytools.js');
 var buildingsExtractor = require('../helpers/buildingsextractor.js');
 var addressesExtractor = require('../helpers/addressextractor.js');
 var publicSpacesExtractor = require('../helpers/publicspacesextractor.js');
@@ -89,92 +88,20 @@ describe('histograph-data-bag', function describeTests() {
   });
 
   describe('conversion phase', function () {
-    describe('geometry checking functions', function () {
-      it('should invalidate an invalid feature', () => {
-        var invalidFeature = {
-          type: 'Feature',
-          properties: { name: 'My non-simple hourglass-shaped geometry' },
-          geometry: {
-            type: 'Polygon',
-            coordinates: [
-              [
-                [5.6, 52.4],
-                [6.3, 52.9],
-                [6.8, 52.1],
-                [7.2, 52.6],
-                [5.6, 52.4]
-              ]
-            ]
-          }
-        };
+    var jobs;
 
-        return geometryTools.validateCoords(invalidFeature.geometry.coordinates, invalidFeature.geometry.type)
-          .then(valid => expect(valid).to.be.false)
-          .catch(errs => {
-            console.error('Validation errors:', errs);
-            return expect(errs).to.be.not.null;
-          });
+    before('create jobs object', () => {
+      jobs = bag.mapFilesToJobs('./test/unzip', './test');
+    });
+
+    describe('file name mapping to jobs', () => {
+      it('should map the files to a list of jobs', done => {
+        expect(jobs.length).to.equal(7);
+        expect(jobs[0].inputFile.split('.').slice(-1)[0]).to.deep.equal('xml');
+        expect(jobs[0].outputPITsFile.split('.').slice(-2)[0]).to.deep.equal('pits');
+        expect(jobs[0].outputRelationsFile.split('.').slice(-2)[0]).to.deep.equal('relations');
+        done();
       });
-
-      it('should reproject the coordinates to WGS84', () => {
-        var geojson = {
-          uri: 'http://bag.kadaster.nl/pand/0362100100084298',
-          id: '0362100100084298',
-          bouwjaar: '2011',
-          geometry: {
-            type: 'Polygon',
-            coordinates: [
-              [
-                [
-                  117283.951,
-                  475941.101
-                ],
-                [
-                  117284.742,
-                  475944.408
-                ],
-                [
-                  117280.949,
-                  475945.315
-                ],
-                [
-                  117280.344,
-                  475942.787
-                ],
-                [
-                  117283.951,
-                  475941.101
-                ]
-              ]
-            ]
-          }
-        };
-
-        expect(geometryTools.toWGS84(geojson.geometry.coordinates[0][0])).to.deep.equal([4.834646702778442, 52.27019375226181]);
-
-      });
-
-      it('should join a gml-extracted position list to a WGS84 geojson-compatible one', () => {
-        var testPosList = '116938.595 477068.148 0.0 ' +
-          '116930.644 477071.854 0.0 ' +
-          '116928.365 477066.959 0.0 ' +
-          '116936.316 477063.253 0.0 ' +
-          '116936.327 477063.277 0.0 ' +
-          '116938.595 477068.148 0.0';
-
-        return geometryTools.joinGMLposlist(testPosList, 'Polygon')
-          .then(geojsoncoords => {
-            console.log(JSON.stringify(geojsoncoords, null, 2));
-            return geometryTools.validateCoords(geojsoncoords, 'Polygon')
-              .then(valid => expect(valid).to.be.true)
-              .catch(err => {
-                console.log('geometry validation error:', err.stack);
-                return expect(err).to.be.null;
-              });
-          });
-
-      });
-
     });
 
     describe('buildings extraction', function() {
@@ -221,33 +148,9 @@ describe('histograph-data-bag', function describeTests() {
         });
       });
 
-      it('should find the building files', () => {
-        var buildingfiles = bag.listBuildingFiles(path.join(__dirname, 'unzip'));
-        console.log('Files:', buildingfiles);
-        return expect(buildingfiles).to.deep.equal(
-          [
-            path.join(__dirname, 'unzip', '1050PND08032011-01022011-0001.xml'),
-            path.join(__dirname, 'unzip', '1050PND08032011-01022011-0002.xml'),
-            path.join(__dirname, 'unzip', '1050PND08032011-01022011-0003.xml')
-          ]
-        );
-      });
-
     });
 
     describe('addresses extraction', () => {
-      it('should find the addresses files', () => {
-        var addressfiles = bag.listAddressFiles(path.join(__dirname, 'unzip'));
-        console.log('Files:', addressfiles);
-        return expect(addressfiles).to.deep.equal(
-          [
-            path.join(__dirname, 'unzip', '1050NUM08032011-01022011-0001.xml'),
-            path.join(__dirname, 'unzip', '1050NUM08032011-01022011-0002.xml'),
-            path.join(__dirname, 'unzip', '1050NUM08032011-01022011-0003.xml')
-          ]
-        );
-      });
-
       it('should extract an address from a mocked snippet', done => {
         addressesExtractor.extractFromFile(path.join(__dirname, 'mockups', 'bag-NUM-snippet.xml'), (err, addressNodes, addressEdges) => {
           if (err) throw err;
@@ -276,16 +179,6 @@ describe('histograph-data-bag', function describeTests() {
     });
 
     describe('public spaces extraction', function() {
-      it('should find the public spaces files', () => {
-        var addressfiles = bag.listPublicSpacesFiles(path.join(__dirname, 'unzip'));
-        console.log('Files:', addressfiles);
-        return expect(addressfiles).to.deep.equal(
-          [
-            path.join(__dirname, 'unzip', '1050OPR08032011-01022011.xml')
-          ]
-        );
-      });
-
       it('should extract the public spaces from the snippet', done => {
         publicSpacesExtractor.extractFromFile(path.join(__dirname, 'mockups', 'bag-OPR-snippet.xml'), (err, publicSpaceNodes, publicSpaceEdges) => {
           if (err) throw err;
